+++ nf_core/modules/pipeline_modules.py
@@ -33,111 +33,9 @@ import sys
 
 import nf_core.utils
 from .module_utils import create_modules_json
+from .modules_repo import ModulesRepo
 
 log = logging.getLogger(__name__)
-
-
-class ModulesRepo(object):
-    """
-    An object to store details about the repository being used for modules.
-
-    Used by the `nf-core modules` top-level command with -r and -b flags,
-    so that this can be used in the same way by all sucommands.
-    """
-
-    def __init__(self, repo="nf-core/modules", branch="master"):
-        self.name = repo
-        self.branch = branch
-        self.modules_file_tree = {}
-        self.modules_current_hash = None
-        self.modules_avail_module_names = []
-
-    def get_modules_file_tree(self):
-        """
-        Fetch the file list from the repo, using the GitHub API
-
-        Sets self.modules_file_tree
-             self.modules_current_hash
-             self.modules_avail_module_names
-        """
-        api_url = "https://api.github.com/repos/{}/git/trees/{}?recursive=1".format(self.name, self.branch)
-        r = requests.get(api_url, auth=nf_core.utils.github_api_auto_auth())
-        if r.status_code == 404:
-            log.error("Repository / branch not found: {} ({})\n{}".format(self.name, self.branch, api_url))
-            sys.exit(1)
-        elif r.status_code != 200:
-            raise SystemError(
-                "Could not fetch {} ({}) tree: {}\n{}".format(self.name, self.branch, r.status_code, api_url)
-            )
-
-        result = r.json()
-        assert result["truncated"] == False
-
-        self.modules_current_hash = result["sha"]
-        self.modules_file_tree = result["tree"]
-        for f in result["tree"]:
-            if f["path"].startswith("software/") and f["path"].endswith("/main.nf") and "/test/" not in f["path"]:
-                # remove software/ and /main.nf
-                self.modules_avail_module_names.append(f["path"][9:-8])
-
-    def get_module_file_urls(self, module):
-        """Fetch list of URLs for a specific module
-
-        Takes the name of a module and iterates over the GitHub repo file tree.
-        Loops over items that are prefixed with the path 'software/<module_name>' and ignores
-        anything that's not a blob. Also ignores the test/ subfolder.
-
-        Returns a dictionary with keys as filenames and values as GitHub API URIs.
-        These can be used to then download file contents.
-
-        Args:
-            module (string): Name of module for which to fetch a set of URLs
-
-        Returns:
-            dict: Set of files and associated URLs as follows:
-
-            {
-                'software/fastqc/main.nf': 'https://api.github.com/repos/nf-core/modules/git/blobs/65ba598119206a2b851b86a9b5880b5476e263c3',
-                'software/fastqc/meta.yml': 'https://api.github.com/repos/nf-core/modules/git/blobs/0d5afc23ba44d44a805c35902febc0a382b17651'
-            }
-        """
-        results = {}
-        for f in self.modules_file_tree:
-            if not f["path"].startswith("software/{}".format(module)):
-                continue
-            if f["type"] != "blob":
-                continue
-            if "/test/" in f["path"]:
-                continue
-            results[f["path"]] = f["url"]
-        return results
-
-    def download_gh_file(self, dl_filename, api_url):
-        """Download a file from GitHub using the GitHub API
-
-        Args:
-            dl_filename (string): Path to save file to
-            api_url (string): GitHub API URL for file
-
-        Raises:
-            If a problem, raises an error
-        """
-
-        # Make target directory if it doesn't already exist
-        dl_directory = os.path.dirname(dl_filename)
-        if not os.path.exists(dl_directory):
-            os.makedirs(dl_directory)
-
-        # Call the GitHub API
-        r = requests.get(api_url, auth=nf_core.utils.github_api_auto_auth())
-        if r.status_code != 200:
-            raise SystemError("Could not fetch {} file: {}\n {}".format(self.name, r.status_code, api_url))
-        result = r.json()
-        file_contents = base64.b64decode(result["content"])
-
-        # Write the file contents
-        with open(dl_filename, "wb") as fh:
-            fh.write(file_contents)
 
 
 class PipelineModules(object):
